requirements.txt
# requirements.txt
Flask==2.3.3
rembg==2.0.50
Pillow==10.0.1
numpy==1.24.3
